backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=10.2.0
rapidfuzz>=3.6.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
aiosqlite>=0.19.0

# OCR dependencies
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0

